+++ template/karma.unit.js
@@ -21,7 +21,10 @@ module.exports = function(config) {
     colors: true,
     logLevel: config.LOG_INFO,
     autoWatch: false,
-    browsers: ['PhantomJS'],
+    browsers: ['Chrome'],
+    mime: {
+      'text/x-typescript': ['ts']
+    },
     singleRun: true
   });
 };